# swagger/two_way_sync.yaml
openapi: 3.0.3
info:
  title: go-oversync Two-Way Sync API
  version: 1.0.0
  description: |
    Bundle-based two-way sync API for single-user multi-device replication.

    The server treats business tables as authoritative. Replication history is derived from
    committed business-table effects and exposed as complete committed bundles. Registered
    PostgreSQL tables are scope-bound on the server, but the hidden server owner column does not
    appear on the wire.
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /sync/connect:
    post:
      summary: Resolve first-connect account attachment
      description: |
        Resolves whether the authenticated user's sync scope is already authoritative on the
        server, whether this source may initialize it from local pending rows, whether the server
        may initialize it empty, or whether the client should retry later because another
        initializer currently owns the scope. `retry_later` is a normal lifecycle outcome, not an
        authentication failure.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectRequest'
      responses:
        '200':
          description: Connect lifecycle resolved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectResponse'
        '400':
          description: Invalid connect request (`connect_invalid`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/push-sessions:
    post:
      summary: Create or restart chunked push session
      description: |
        Creates or restarts the transport session for one logical dirty-set bundle identified by
        `(source_id, source_bundle_id)`. If that tuple was already committed previously, the server
        returns `already_committed` plus the authoritative committed bundle metadata instead of
        creating a new staging session. If the tuple is already known but retained replay metadata
        for that tuple has been pruned, the server returns `409 history_pruned` instead of
        reopening staging. If `source_bundle_id` is not the next expected contiguous value for the
        current source, the server returns `409 source_sequence_out_of_order`. If the current
        source was explicitly retired during rotated rebuild, the server returns
        `409 source_retired` with replacement lineage.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSessionCreateRequest'
      responses:
        '200':
          description: Push session created or already-committed result returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushSessionCreateResponse'
        '400':
          description: Invalid push-session request (`push_session_invalid`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Scope not ready, retained replay for the requested source tuple is pruned, the source bundle sequence is not the next expected value, or the source was explicitly retired (`scope_uninitialized`, `scope_initializing`, `initialization_stale`, `history_pruned`, `source_sequence_out_of_order`, or `source_retired`)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/SourceRetiredResponse'
        '410':
          description: Initialization lease expired before push create (`initialization_expired`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/push-sessions/{push_id}/chunks:
    post:
      summary: Upload one ordered push chunk
      description: |
        Uploads one ordered chunk for a staging push session. `rows[]` reuse the `PushRequestRow`
        contract exactly. `start_row_ordinal` must equal the next expected ordinal for the session.
        Empty chunks are rejected. Chunks whose row count exceeds the server-advertised
        `max_rows_per_push_chunk` are rejected fail-closed.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
        - name: push_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CanonicalSessionToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushSessionChunkRequest'
      responses:
        '200':
          description: Push chunk accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushSessionChunkResponse'
        '400':
          description: Invalid chunk upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Out-of-order chunk upload or stale initialization lease (`push_chunk_out_of_order` or `initialization_stale`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '403':
          description: Push session belongs to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Push session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Push session expired or initialization lease expired (`push_session_expired` or `initialization_expired`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/push-sessions/{push_id}/commit:
    post:
      summary: Commit one staged push session
      description: |
        Validates the full staged dirty set, runs whole-bundle conflict detection, and commits the
        authoritative business-table effects atomically. On success, the response returns the
        authoritative committed bundle metadata; the client fetches committed rows separately. If
        the per-source bundle sequence changes between create-time validation and commit-time
        finalization, the server returns `409 source_sequence_changed`. If the source was retired
        after staging began, the server returns `409 source_retired`.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
        - name: push_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CanonicalSessionToken'
      responses:
        '200':
          description: Push session committed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushSessionCommitResponse'
        '400':
          description: Invalid commit request or incomplete staged rows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '403':
          description: Push session belongs to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Push session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Whole-bundle conflict rejection, stale initialization lease, commit-time source sequence revalidation failure, or explicit source retirement (`push_conflict`, `initialization_stale`, `source_sequence_changed`, or `source_retired`)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/PushConflictResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/SourceRetiredResponse'
        '410':
          description: Push session expired or initialization lease expired (`push_session_expired` or `initialization_expired`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/push-sessions/{push_id}:
    delete:
      summary: Delete uncommitted push session
      description: Best-effort explicit cleanup of an uncommitted staging push session.
      security:
        - BearerAuth: []
      parameters:
        - name: push_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/CanonicalSessionToken'
      responses:
        '204':
          description: Push session deleted successfully
        '400':
          description: Invalid delete request (`push_chunk_invalid`, for example when the push session is already committed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '403':
          description: Push session belongs to a different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Push session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Push session expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/committed-bundles/{bundle_seq}/rows:
    get:
      summary: Fetch authoritative committed bundle rows
      description: |
        Returns one deterministic page of the authoritative committed bundle row stream. `rows[]`
        reuse the `BundleRow` contract exactly. Clients use this endpoint for authoritative replay
        after chunked push commit and for post-commit crash recovery. Omit `after_row_ordinal` on
        the first request; when provided, the server returns rows strictly after that ordinal.
        For `bundle_hash` verification, clients reconstruct absolute `row_ordinal` from stream
        position: the first row of the first page is ordinal 0, the first row after
        `after_row_ordinal = N` is ordinal `N + 1`, and each following row increments by 1 in order.
        Requests for bundles at or below the retained history floor return `409 history_pruned`
        even if those rows have not yet been physically deleted.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
        - name: bundle_seq
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: after_row_ordinal
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
        - name: max_rows
          in: query
          required: false
          description: Optional requested committed-bundle chunk size. Defaults to 1000, is capped by the server at 5000, and non-positive values are rejected.
          schema:
            type: integer
            minimum: 1
            default: 1000
            maximum: 5000
      responses:
        '200':
          description: Committed bundle row page retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommittedBundleRowsResponse'
        '400':
          description: Invalid committed-bundle row request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '409':
          description: Requested committed bundle is below the retained history floor (`history_pruned`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Bundle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/pull:
    get:
      summary: Pull committed bundles
      description: |
        Returns complete committed bundles only. The first response freezes a stable pull
        ceiling (`stable_bundle_seq`). Follow-up requests should pass the same
        `target_bundle_seq` until the client reaches that ceiling. The retained bundle floor is
        the highest discarded bundle: a checkpoint equal to the floor is valid and the response
        contains only bundles strictly above it. A positive floor rejects lower checkpoints,
        including zero. Checkpoints and positive targets above current committed history are
        rejected rather than clamped.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
        - name: after_bundle_seq
          in: query
          required: false
          description: Pull bundles strictly after this durable client checkpoint.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
        - name: max_bundles
          in: query
          required: false
          description: Maximum number of complete bundles to return.
          schema:
            type: integer
            minimum: 1
            maximum: 5000
            default: 1000
        - name: target_bundle_seq
          in: query
          required: false
          description: Frozen stable ceiling captured from the first response in a pull session. Zero asks the server to capture the current committed ceiling.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: Pull page retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullResponse'
        '400':
          description: Invalid pull parameters, including a positive target below `after_bundle_seq` (`invalid_request`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '409':
          description: Requested checkpoint/target is below retained history or ahead of current committed history, or the scope is not initialized (`history_pruned`, `checkpoint_ahead`, `scope_uninitialized`, or `scope_initializing`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service unavailable during graceful shutdown (`service_unavailable`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
  /sync/watch:
    get:
      summary: Watch for committed bundle wakeups
      description: |
        Opens a Server-Sent Events stream that emits metadata-only wake-up hints when the
        authenticated scope has newer committed bundles. The stream is not a data delivery path:
        clients must call `/sync/pull` or existing sync operations to fetch authoritative data.
        Heartbeat comments keep idle connections observable and allow stale subscribers to be
        cleaned up.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
        - name: after_bundle_seq
          in: query
          required: false
          description: Last durable bundle checkpoint known by the client. Missing means 0.
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
      responses:
        '200':
          description: SSE stream of `bundle` events and heartbeat comments.
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                bundle:
                  value: |
                    event: bundle
                    data: {"bundle_seq":124,"source_id":"device-a","source_bundle_id":7}

                    : heartbeat

        '400':
          description: Invalid `after_bundle_seq` (`invalid_request`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '403':
          description: Watch is disabled for this authenticated actor/source (`bundle_change_watch_forbidden`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Scope is uninitialized or initializing (`scope_uninitialized` or `scope_initializing`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Bundle change watch is disabled (`bundle_change_watch_disabled`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected watch setup failure (`bundle_change_watch_failed`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /sync/capabilities:
    get:
      summary: Get protocol capabilities
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SourceIDHeader'
      responses:
        '200':
          description: Capabilities retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilitiesResponse'
        '401':
          description: Authentication required
components:
  parameters:
    SourceIDHeader:
      name: Oversync-Source-ID
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/SourceID'
      description: |
        Current sync source identity for this authenticated request. This header is sync transport
        metadata, not authentication. Direct protocol callers provide it explicitly; higher-level
        clients such as oversqlite may manage it internally.
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PushSessionCreateRequest:
      type: object
      required: [source_bundle_id, planned_row_count, canonical_request_hash]
      properties:
        source_bundle_id:
          type: integer
          format: int64
          minimum: 1
        planned_row_count:
          type: integer
          format: int64
          minimum: 1
        canonical_request_hash:
          $ref: '#/components/schemas/CanonicalRequestHash'
        initialization_id:
          $ref: '#/components/schemas/CanonicalSessionToken'
    ConnectRequest:
      type: object
      required: [has_local_pending_rows]
      properties:
        has_local_pending_rows:
          type: boolean
          description: Whether this client captured anonymous local rows that could seed an uninitialized scope.
    ConnectResponse:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string
          enum: [remote_authoritative, initialize_local, initialize_empty, retry_later]
        initialization_id:
          $ref: '#/components/schemas/CanonicalSessionToken'
        lease_expires_at:
          type: string
          format: date-time
          description: Lease expiry for `initialize_local` and `retry_later` responses when the server currently tracks an initializer.
        retry_after_seconds:
          type: integer
          minimum: 1
          description: Suggested retry backoff when `resolution = retry_later`.
    PushSessionCreateResponse:
      oneOf:
        - $ref: '#/components/schemas/PushSessionCreateStagingResponse'
        - $ref: '#/components/schemas/PushSessionCreateAlreadyCommittedResponse'
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        required_byte_count:
          type: integer
          format: int64
...
    SourceRetiredResponse:
      type: object
      required: [error, message, source_id]
      properties:
        error:
          type: string
          enum: [source_retired]
        message:
          type: string
        source_id:
          allOf:
            - $ref: '#/components/schemas/SourceID'
          description: Exact source ID used by the failed request. A client must reject the response unless this value exactly matches its request source before changing recovery state.
        replaced_by_source_id:
          allOf:
            - $ref: '#/components/schemas/SourceID'
          description: Optional by omission only and may be omitted when no replacement is known. When present, this is the non-empty replacement source reserved or activated for the retired source. Explicit null and empty values are malformed.
...